backend/requirementsDjango.txt
appdirs==1.4.4
asgiref==3.3.1
distlib==0.3.1
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
importlib-metadata==3.4.0
mysql-connector-python==8.0.23
protobuf==3.14.0
psycopg2==2.8.6
psycopg2-binary
PyJWT==2.0.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
virtualenv==20.4.2
zipp==3.4.0
